[ts-sdk] mint and set ans subdomain
Subdomains do not by default set the target, so we have two calls... one
to mint and another to set...

note there was a bug in the SDK where it made an assumption that targets
are always set, this was fixed.

there was also another bug around handling of re-minting after
expiration, which has also been fixed.