[Sharded-Executor]: Add support for 'total_supply' in sharded executor. (#9882)
* [Sharded-Executor]: Add support for 'total_supply' in sharded executor.

We achieve this by aggregating the total_supply deltas from individual shards
and then updating the txns on all shards with the correct total supply.

Since an individual shard can have a negative delta for 'total_supply' (say due to burn), we start with the base value of u128/2 on every shard. We collect the deltas from u128/2 and then in the second pass update the every txn with its delta over the actual base value.

For now we update the write set of every txn 'in-place' with the final value of total_supply to avoid deserializing and serializing the entire write set to update the total_supply with correct aggregated delta.