[framework] Turn on test-only checking and fix errors (#10368)
This turns on running extended checks also on test-only code in the framework unit tests. A few errors discovered this way are fixed.

Changes to function declarations in this PR do not effect compatibility because only test-only functions are effected which are stripped before deployment.

Related to #10335, but more needs to be done to make this behavior the default. This cannot happen before the next framework release.