[consensus] reduce block size to 2500, simplify logic by removing quorum store overrides (#11880)
### Description

With the improvements in VM load time, it's no longer necessary to have a very large block size for throughput. In fact, with the block gas limits, a large block size can be bad, because it leads to frequent block cuts.

With targeting a smaller block size, there's no more reason to have a separate quorum store enabled size. So combined into a single block size and simplified the logic.

### Test Plan

Run e2e perf test.